library/requirements.txt
art==5.6
numpy==1.22.3
Pillow==9.1.0
